@@ -5,7 +5,6 @@ import 'package:flutter/services.dart';
 import 'package:get/get.dart';
 import 'package:orange_ui/api_provider/api_provider.dart';
 import 'package:orange_ui/common/widgets/common_ui.dart';
-import 'package:orange_ui/common/widgets/snack_bar_widget.dart';
 import 'package:orange_ui/generated/l10n.dart';
 import 'package:orange_ui/model/get_diamond_pack.dart';
 import 'package:stacked/stacked.dart';
@@ -21,7 +20,7 @@ class BottomDiamondShopViewModel extends BaseViewModel {
       if (call.arguments == true) {
         addCoinApiCall();
       } else {
-        SnackBarWidget.snackBar(message: S.current.failedPayment);
+        CommonUI.snackBar(message: S.current.failedPayment);
       }
       return;
     });
