@@ -45,7 +45,7 @@ class LikeCard extends StatelessWidget {
                 fit: BoxFit.cover,
                 errorWidget: (context, url, error) {
                   return CommonUI.profileImagePlaceHolder(
-                      name: userData.fullname, heightWeight: 40);
+                      name: userData.fullname, heightWidth: 40);
                 },
               ),
             ),
